# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /novels/{id}/chapters:
    get:
      summary: Get chapters of a novel
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chapter'
...
  /chapters:
    post:
      summary: Create new chapter (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, title]
              properties:
                novel_id:
                  type: integer
                title:
                  type: string
                order_index:
                  type: integer
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chapters/{id}:
    put:
      summary: Update chapter (Writer)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                order_index:
                  type: integer
      responses:
        '200':
          description: Chapter updated successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete chapter (Writer)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chapter deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource Not Found Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    Chapter:
      type: object
      properties:
        id:
          type: integer
        novel_id:
          type: integer
        title:
          type: string
        order_index:
          type: integer
...